Extract out common subset helper
