+++ crates/ark/src/lsp/completions/sources/common.rs
@@ -1,0 +1,8 @@
+//
+// common.rs
+//
+// Copyright (C) 2024 Posit Software, PBC. All rights reserved.
+//
+//
+
+pub(crate) mod subset;
