+++ crates/ark/src/lsp/completions/sources/composite/subset.rs
@@ -7,12 +7,10 @@
 
 use anyhow::Result;
 use tower_lsp::lsp_types::CompletionItem;
-use tree_sitter::Node;
-use tree_sitter::Point;
 
+use crate::lsp::completions::sources::common::subset::is_within_subset_delimiters;
 use crate::lsp::completions::sources::utils::completions_from_evaluated_object_names;
 use crate::lsp::document_context::DocumentContext;
-use crate::lsp::traits::point::PointExt;
 use crate::lsp::traits::rope::RopeExt;
 use crate::treesitter::NodeType;
 use crate::treesitter::NodeTypeExt;
@@ -71,50 +69,6 @@ pub(super) fn completions_from_subset(
     let text = context.document.contents.node_slice(&child)?.to_string();
 
     completions_from_evaluated_object_names(&text, ENQUOTE)
-}
-
-fn is_within_subset_delimiters(x: &Point, subset_node: &Node, subset_type: &NodeType) -> bool {
-    let (open, close) = match subset_type {
-        NodeType::Subset => ("[", "]"),
-        NodeType::Subset2 => ("[[", "]]"),
-        _ => std::unreachable!(),
-    };
-
-    let Some(arguments) = subset_node.child_by_field_name("arguments") else {
-        return false;
-    };
-
-    let n_children = arguments.child_count();
-
-    if n_children < 2 {
-        return false;
-    }
-
-    let Some(open_node) = arguments.child(1 - 1) else {
-        return false;
-    };
-    let Some(close_node) = arguments.child(n_children - 1) else {
-        return false;
-    };
-
-    // Ensure open and closing nodes are the right type
-    if !matches!(
-        open_node.node_type(),
-        NodeType::Anonymous(kind) if kind == open
-    ) {
-        return false;
-    }
-    if !matches!(
-        close_node.node_type(),
-        NodeType::Anonymous(kind) if kind == close
-    ) {
-        return false;
-    }
-
-    let contains_start = x.is_after_or_equal(open_node.end_position());
-    let contains_end = x.is_before_or_equal(close_node.start_position());
-
-    contains_start && contains_end
 }
 
 #[cfg(test)]
